Adjust scanner branch conditions and branch types
For RISC-V the `cond` field is a `mambo_cond` struct to store both
compared registers and the condition type in a more readable form
than (mis-)using the `rn` field for that ;). (The `rn` field is unused for
conditional branches)

NOTE: This lets the code cache meta array take up a little more space.