Fix `ra` register pushed before it is changed
On context switch the `ra` register must be pushed before any
instruction writing to it. This includes the call to the subroutines
that push other registers like `push_x1_x31`. Of course, `ra` needs to
be restored between the last instruction changing `ra` and the execution
of instrumented client code.