Fix branch offset for branching directly to next BB
Every instrumented basic block starts with popping x10 and x11 from the
stack but this should only be executed if the dispatcher was invoked
before. In a situation like an unconditional direkt jump, the jump is
instrumented to a jump directly to the next basic block in the code cache.
No dispatcher called in that case, so x10 and x11 are not pushed and
should not be popped.