Improve LR/SC translation performance
Since the LR instruction is the start of the software emulation of the
LR/SC loop and the atomic hardware instructions are inserted in the
translation of SC, the LR instruction can be translated to a regular LD
or LW instruction. This speeds up the execution of the translated
sequence by a lot!