Rework makefile for better scaling
The makefile now supports C, Assembler and C++ (new) source files. It
builds seperate object files in a build directory and eventually links
them together to one executable. The type of the compile needed to build
a source file is determined by it's suffix:
	- AS:	.s
	- CC: 	.c, .S
	- CXX: 	.cpp

This allows MAMBO and it's plugins to include C++ libraries.