Fix `emit_fcall` to make context switch of gp and tp
When a plugin function call is inserted with `emit_fcall` and it uses
static variables of the plugin or DBM engine, the global pointer and the
thread pointer must have the values of the DBM engine context despite
the code running in the client context. It is needed because static
variables that are defined in a plugin or the engine are compiled to the
MAMBO context.