Add FIXME note to scanner post basic block delivery
When post basic block and post fragment callbacks are delivered, the
`read_address` field of the code context doesn't correspond to the
`write_p` field. `read_address` still points to the beginning of the
basic block in the original scanned code, whereas `write_p` points to
the end of the translated basic block (as expected).