Add full encryption run to `InitTarget`
Due to lazy binding of the PLT entries, all external functions in the
target code should be called at least once to resolve the PLT bindings.
The first call to a PLT entry will invoke the dynamic linker but all
subsequent calls to the same PLT entry are linked directly and behave in
the same way.