Add recognition of calls and returns in tracer
Microwalk analysis wants to know when a function is called and when
it returns. In RISC-V we don't have "call" and "ret" instructions like on
x86(_64), but call and return operations are usually translated into one of
the following instructions:

A call is a jump and link instruction, where the link register is x1:
	JAL     x1, <imm>
	JALR    x1, <rs1>, <imm>
	C.JAL   <imm>
	C.JALR  <rs1>

A return is an indirect jump instruction, where the register is x1 and the link
register x0:
	JALR    x0, x1, <imm>
	C.JR    x1

But this is not guaranteed, as these are not the only ways to translate call
and return operations. For example, identical behavior can be constructed with
conditional branches:
	AUIPC   x1, 0
	ADDI    x1, x1, 12
	BEQ     x0, x0, <offset>

or by simply not following the standard calling convention and using another
register than x1 as the link register like so:
	JALR    x10, x11, <imm>
Btw, the above example could also be an obfuscated return operation.