Whitelist passed by file 'whitelist.txt'
The whitelist is now passed by file replacing the hacky solution using
a command line argument. This way, we don't interfere with the arguments
passed to the target.

The tracer plugin expects a file named 'whitelist.txt' at the current
working directory, otherwise it will complain.

Example:
Example cwd:
.
├── mambo-v/
│   └── dbm
└── whitelist.txt

Run it with:
$ mambo-v/dbm /path/to/fancy/program